fix: add fake DynamicClient, move NoOp to repositories
- Add fakedynamic.NewSimpleDynamicClient to NewFakeAPIServerStorage
- Move NoOpSecurityExceptionRepository from adapters/v1 to repositories
