+++ repositories/noop_security_exception.go
@@ -1,4 +1,4 @@
-package v1
+package repositories
 
 import (
 	"context"
